Refactor plots utilities into visualization modules
